+++ flutter_project/lib/ptvInfoClasses/departure_info.dart
